Make base class AppNotificationSaver
